♻️ refactor: 서버 함수 폴더 구조 변경에 따른 경로 재설정
